fix aiohttp_cors to work with aiohttp==0.21
This commit adds workaround for missing "match()" method, see
<https://github.com/KeepSafe/aiohttp/issues/766>.

Adds support in RouterAdapter for Resource-based routes.

However configuring CORS for Resources is not yet supported.

This change will make pre-0.21 Route-based user code to work with
aiohttp 0.21, if user doesn't used name-based lookup of routes
(which is one of backward incompatible changes in 0.21).